policy: Reverse locking order between selector cache and name manager
The locking order between the NameManager lock and the SelectorCache
lock has been accidental so far. When removing a selector user from
the cache, the SelectorCache lock is taken, and if an FQDN selector
has no more users, the NameManager is notified of this, which
internally takes the NameManager lock. Change this by deferring the
NamaManager callback from the selector cache until the selector cache
mutex is unlocked.

From now on, if the two locks are to be held at the same time, the
NameManager mutex MUST be taken first, the SelectorCache mutex second,
as the name manager is a higher level construct than the selector
cache.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>