fqdn: Keep name manager locked during selector cache updates
UpdateGenerateDNS updated the cache and generate selector updates as two
separate lock-holding sections. When updates happened concurrently, via
DNS lookups, DNS GC, or endpoint restores, the snapshot of data that
eventually updated the selectors could be out-of-order. This created an
A-B-A race for selectors that needed an update and those that needed to
be cleared.

These races have been observed on agent starts with pending pods but may
manifest between multiple pods making DNS updates or the DNS garbage
collector controller and any pod.

The code now holds the NameManager lock for the entirety of
UpdateGenerateDNS and ForceRegenerateDNS. This forces the update to
complete in its entirety before allowing another to modify the FQDN
cache or registered selectors.

This more aggressive locking depends on the locking order allowing
selector cache locking while holding name manager lock ewhich wass
introduced in the previous commit.

Signed-off-by: Ray Bejjani <ray@isovalent.com>
Signed-off-by: Jarno Rajahalme <jarno@covalent.io>