Replace duration between events metric with events_ts.
The events_ts metric already handles k8s events, this
adds scope and action labels to the events_ts metric.
To accomade this I removed the duplicate declarations
of the events_ts metrics that used constant labels and
replaced it with a single GaugeVector metric type.
As a result, the API event metrics will be tagged with
additional labels.
ContainerD metric appeared not to be used.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>