bpf: add optional map event log buffer to bpf.Map.
Add implementation for ring buffer based map events log.
Map events store a timestamped event log object for every attempted
operation of a bpf.Map.

Map events that do not sync (such as in the case of syscall errors) are
added to the buffer with the relevant sync state.
All subsequent attempts reconcile the bpf map state via the controller
will also be added to the event log buffer.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>